python/emqx_example/pyproject.toml
[project]
name = "emqx-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["paho-mqtt>=2.0.0"]
